+++ imio/dms/mail/examples.py
@@ -1030,8 +1030,6 @@ def configure_imio_dms_mail(context):
         registry["imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_types"] = [
             {"value": u"type1", "dtitle": u"Type 1", "active": True},
         ]
-    if not registry.get("imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_odt_mainfile"):
-        registry["imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_odt_mainfile"] = True
     if not registry.get("imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_response_prefix"):
         registry["imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_response_prefix"] = _(u"Response: ")
     if not registry.get("imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_send_modes"):
@@ -1042,6 +1040,8 @@ def configure_imio_dms_mail(context):
         ]
     if registry.get("imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_replyto_email_send") is None:
         registry["imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_replyto_email_send"] = False
+    if registry.get("imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_formats_mainfile") is None:
+        registry["imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_formats_mainfile"] = ["odt"]
     if not registry.get("imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_fields"):
         fields = [
             "IDublinCore.title",