+++ imio/dms/mail/utils.py
@@ -1205,7 +1205,7 @@ class OdmUtilsMethods(UtilsMethods):
 
     def is_odt_activated(self):
         registry = getUtility(IRegistry)
-        return registry["imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_odt_mainfile"]
+        return ["odt"] == registry["imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_formats_mainfile"]
 
     def scanned_col_cond(self):
         """Condition for searchfor_scanned collection"""
@@ -1635,3 +1635,17 @@ def update_approvers_settings():
         for userid in removed_approvings:
             s_l.manage_delLocalRoles([userid])
             s_l.reindexObject()
+
+
+def get_allowed_omf_content_types():
+    """Get allowed outgoing mail main file content types."""
+    ct_by_type = {
+        "pdf": ("application/pdf",),
+        "odt": ("application/vnd.oasis.opendocument.text",),
+        "doc": ("application/msword", "application/vnd.openxmlformats-officedocument.wordprocessingml.document"),
+    }
+    formats = api.portal.get_registry_record("imio.dms.mail.browser.settings.IImioDmsMailConfig.omail_formats_mainfile")
+    res = []
+    for fmt in formats or []:
+        res.extend(ct_by_type.get(fmt, ()))
+    return res